Split dropwizard-jetty's ZipExceptionHandlingInputStreamTest so that it gets run
The default surefire configuration doesn't execute tests in static inner
classes. We can't easily make the inner classes non-static because of the
parameterized tests, so just split the test class into two classes instead.
